I think I successsfully made the switch to ints. Attacks have resumed, ransoms being paid. Everything is now in terms of THOUSANDS of dollars. estimates from python script is_attacking_worth_it_exploration.py indicate that integer overflow is extremely unlikely (and should be caught if it ever happens anyway)

+++ cpp-games/src/Attacker.h
@@ -9,10 +9,10 @@ class Attacker : public Player {
 
         static double inequality_ratio;
         
-        static double a_init;
-        static double attacker_iter_sum; // how much the attackers have cumulatively gained or lost this round
-        static std::vector<double> cumulative_assets; // running total of all attackers' assets
-        static double current_sum_assets; // sum total of all class instances
+        static long long  a_init;
+        static long long  attacker_iter_sum; // how much the attackers have cumulatively gained or lost this round
+        static std::vector<unsigned long long> cumulative_assets; // running total of all attackers' assets
+        static long long  current_sum_assets; // sum total of all class instances
 
         static double estimated_current_defender_wealth_mean;
         static double estimated_current_defender_wealth_stdddev;
@@ -20,17 +20,17 @@ class Attacker : public Player {
         static double estimated_current_defender_posture_stdddev;
 
         // bookkeeping vars for all attackers combined
-        static int attacksAttempted;
-        static int attacksSucceeded;
-        static double attackerExpenditures;
-        static double attackerLoots;
+        static long long  attacksAttempted;
+        static long long  attacksSucceeded;
+        static long long  attackerExpenditures;
+        static long long  attackerLoots;
 
 
         static void reset();
 
         Attacker(int id_in,  Params &p);
-        void gain(double gain) override;
-        void lose(double loss) override;
+        void gain(uint32_t gain) override;
+        void lose(uint32_t loss) override;
 
         // To re-estimate current defender params
         void perform_market_analysis(); // TODO TODO TODO implement
